pyproject.toml
[project]
name = "mockup-generator"
version = "0.1.0"
description = "Luxury garment mockup tool"
authors = [{ name = "bindal.abhinav@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "openai>=1.99.3,<2.0.0",
  "pillow>=11.3.0,<12.0.0",
  "python-dotenv (>=1.1.1,<2.0.0)",
  "google-genai (>=1.36.0,<2.0.0)",
  "streamlit (>=1.50.0,<2.0.0)"
]

[project.scripts]
mockup = "mockup_generator.create_mockup:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
